Cleans up Skinning + Morph ( loading / compiler )
RigGeometry has methods to get children vertex and primitives to copy
them into the RigGeometry itself. It's not ideal but it's very simple
and anyway at one time we need to bind everything to be able to animate
it. I thought to build the vertex attributes list dynamically, but it's
not really better if users want to inspect or change data. So the final
method is to get Morph attributes / primitives and copy them to the
RigGeometry.